make step function declaration generic
